requirements.txt
torch>=2.1
torchvision>=0.16
timm>=1.0
pandas>=2.0
pyarrow>=14.0
pillow>=10.0
numpy>=1.26
tqdm>=4.66
